handlers/gunshot_detection/requirements.txt
torchaudio
pyyaml